Hacked together a collapsing/expanding animation in css
So apparently this is not flawlessly possible in pure css because transitions
are not supported on auto/100% value changes. We would either need to used fixed
pixel height changes, unaware of content size, or a variety of hacks. I decided
to animate the content collapsing and expanding, and have the container size
jump to the correct height at the end of the animation, because you get a
gestalt effect of the container moving, even though it is still slightly
jarring.

Of course, the desired effect is entirely possible using JavaScript, but this
seemed like a fun challenge.